Added youtube video
Video added. You can find it searching "dialogVideo" inside the .html file. Dimensions are hard coded because I had  problem with the height of the div. BUG BOUNTY: when you close the div of the video the audio goes on. REWARD: a beer bought from the bangla store in fron of Ruben's house.